Added switch and single export for api & mockApi
