+++ .storybook/src/api/index.ts
@@ -1,0 +1,6 @@
+import * as apiReal from './api';
+import * as apiMock from './apiMock';
+
+export * from './api';
+// Un comment if developing or running with local data
+// export * from './apiMock';
