--- .storybook/src/exampledata/putexampledatahere.json
@@ -1,1 +1,0 @@
-* sjekke grane data
